+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Implement SideNav Channel Search
 * Implement SideNav More Channels Filters
 * Implement SideNav Create channel
+* Add message searching
 
 ## 0.2.1 (2018-01-15)